Fix: GameRendererMixin crash - getFov returns float not double
CallbackInfoReturnable<Double> caused ClassCastException when Mixin called
getReturnValueF(). Corrected to CallbackInfoReturnable<Float>.

+++ common/src/main/java/dev/vercim/handycam/mixin/GameRendererMixin.java
@@ -21,9 +21,9 @@ public abstract class GameRendererMixin {
      */
     @Inject(method = "getFov", at = @At("RETURN"), cancellable = true)
     private void handycam$blockDynamicFov(Camera camera, float partialTick, boolean useFov,
-                                          CallbackInfoReturnable<Double> cir) {
+                                          CallbackInfoReturnable<Float> cir) {
         if (!HandycamConfig.get().enableVanillaFov && useFov) {
-            cir.setReturnValue((double) (int) Minecraft.getInstance().options.fov().get());
+            cir.setReturnValue((float) (int) Minecraft.getInstance().options.fov().get());
         }
     }
 }
